requirements.txt
-i https://pypi.org/simple
aiofiles>=22.1.0 ; python_version >= '3.7' and python_version < '4.0'
anyio>=3.6.2 ; python_full_version >= '3.6.2'
appdirs>=1.4.4
appnope>=0.1.3 ; platform_system >= 'Darwin'
argcomplete>=2.0.0 ; python_version >= '3.6'
argh>=0.26.2
argon2-cffi>=21.3.0 ; python_version >= '3.6'
argon2-cffi-bindings>=21.2.0 ; python_version >= '3.6'
asttokens>=2.1.0
asyncio>=3.4.3
asyncpg>=0.27.0 ; python_full_version >= '3.7.0'
attrs>=22.1.0 ; python_version >= '3.5'
babel>=2.11.0 ; python_version >= '3.6'
backcall>=0.2.0
beautifulsoup4>=4.11.1 ; python_full_version >= '3.6.0'
biocommons.seqrepo>=0.6.5
biopython>=1.80
bioutils>=0.5.7 ; python_version >= '3.6'
bleach>=5.0.1 ; python_version >= '3.7'
boto3>=1.26.16 ; python_version >= '3.7'
botocore>=1.29.16 ; python_version >= '3.7'
bs4>=0.0.1
canonicaljson>=1.6.4 ; python_version >= '3.7'
certifi>=2022.9.24 ; python_version >= '3.6'
cffi>=1.15.1
charset-normalizer>=2.1.1 ; python_full_version >= '3.6.0'
click>=8.1.3 ; python_version >= '3.7'
coloredlogs>=15.0.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
configparser>=5.3.0 ; python_version >= '3.7'
cool-seq-tool>=0.1.4
cssselect>=1.2.0 ; python_version >= '3.7'
debugpy>=1.6.3 ; python_version >= '3.7'
decorator>=5.1.1 ; python_version >= '3.5'
defusedxml>=0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
entrypoints>=0.4 ; python_version >= '3.6'
executing>=1.2.0
fake-useragent>=1.0.1
fastapi>=0.87.0 ; python_version >= '3.7'
fastjsonschema>=2.16.2
ga4gh.vrs[extras]>=0.8.7.dev0 ; python_version >= '3.6'
ga4gh.vrsatile.pydantic>=0.0.13
gene-normalizer>=0.1.37
gffutils>=0.11.1
h11>=0.14.0 ; python_version >= '3.7'
hgvs>=1.5.4 ; python_version >= '3.6'
humanfriendly>=10.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna>=3.4 ; python_version >= '3.5'
importlib-metadata>=5.0.0 ; python_version >= '3.7'
inflection>=0.5.1 ; python_version >= '3.5'
ipykernel>=6.17.1 ; python_version >= '3.8'
ipython>=8.6.0 ; python_version >= '3.8'
ipython-genutils>=0.2.0
ipywidgets>=8.0.2 ; python_version >= '3.7'
jedi>=0.18.2 ; python_version >= '3.6'
jinja2>=3.1.2 ; python_version >= '3.7'
jmespath>=1.0.1 ; python_version >= '3.7'
json5>=0.9.10
jsonschema>=3.2.0
jupyter>=1.0.0
jupyter-client>=7.4.7 ; python_version >= '3.7'
jupyter-console>=6.4.4 ; python_version >= '3.7'
jupyter-core>=5.0.0 ; python_version >= '3.8'
jupyter-server>=1.23.3 ; python_version >= '3.7'
jupyterlab>=3.5.0
jupyterlab-pygments>=0.2.2 ; python_version >= '3.7'
jupyterlab-server>=2.16.3 ; python_version >= '3.7'
jupyterlab-widgets>=3.0.3 ; python_version >= '3.7'
lxml>=4.9.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markdown>=3.4.1 ; python_version >= '3.7'
markupsafe>=2.1.1 ; python_version >= '3.7'
matplotlib-inline>=0.1.6 ; python_version >= '3.5'
mistune>=2.0.4
nbclassic>=0.4.8 ; python_version >= '3.7'
nbclient>=0.7.0 ; python_full_version >= '3.7.0'
nbconvert>=7.2.5 ; python_version >= '3.7'
nbformat>=5.7.0 ; python_version >= '3.7'
nest-asyncio>=1.5.6
notebook>=6.5.2 ; python_version >= '3.7'
notebook-shim>=0.2.2 ; python_version >= '3.7'
numpy>=1.23.5 ; python_version >= '3.8'
packaging>=21.3 ; python_version >= '3.6'
pandas>=1.5.2
pandocfilters>=1.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parse>=1.19.0
parsley>=1.3
parso>=0.8.3 ; python_version >= '3.6'
pexpect>=4.8.0 ; sys_platform != 'win32'
pickleshare>=0.7.5
platformdirs>=2.5.4 ; python_version >= '3.7'
plotly>=5.11.0
prometheus-client>=0.15.0 ; python_version >= '3.6'
prompt-toolkit>=3.0.33 ; python_full_version >= '3.6.2'
psutil>=5.9.4 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
psycopg>=3.1.9
psycopg2>=2.9.5 ; python_version >= '3.6'
psycopg2-binary>=2.9.5
ptyprocess>=0.7.0 ; os_name != 'nt'
pure-eval>=0.2.2
pycparser>=2.21
pydantic>=1.10.2 ; python_version >= '3.7'
pyee>=8.2.2
pyfaidx>=0.7.1
pygments>=2.13.0 ; python_version >= '3.6'
pyliftover>=0.4
pyparsing>=3.0.9 ; python_full_version >= '3.6.8'
pyppeteer>=1.0.2 ; python_version >= '3.7' and python_version < '4.0'
pyquery>=1.4.3
pyrsistent>=0.19.2 ; python_version >= '3.7'
pysam>=0.18.0
pytest>=6.0
python-dateutil>=2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-jsonschema-objects>=0.4.1
pytz>=2022.6
pyyaml>=6.0 ; python_version >= '3.6'
pyzmq>=24.0.1 ; python_version >= '3.6'
qtconsole>=5.4.0 ; python_version >= '3.7'
qtpy>=2.3.0 ; python_version >= '3.7'
requests>=2.28.1
requests-html>=0.10.0 ; python_full_version >= '3.6.0'
s3transfer>=0.6.0 ; python_version >= '3.7'
send2trash>=1.8.0
setuptools>=65.6.2 ; python_version >= '3.7'
simplejson>=3.18.0 ; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six>=1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio>=1.3.0 ; python_version >= '3.7'
soupsieve>=2.3.2.post1 ; python_version >= '3.6'
sqlparse>=0.4.3 ; python_version >= '3.5'
stack-data>=0.6.1
starlette>=0.21.0 ; python_version >= '3.7'
tabulate>=0.9.0 ; python_version >= '3.7'
tenacity>=8.1.0 ; python_version >= '3.6'
terminado>=0.17.0 ; python_version >= '3.7'
tinycss2>=1.2.1 ; python_version >= '3.7'
tomli>=2.0.1 ; python_version >= '3.7'
tornado>=6.2 ; python_version >= '3.7'
tqdm>=4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
traitlets>=5.5.0 ; python_version >= '3.7'
typing-extensions>=4.4.0 ; python_version >= '3.7'
urllib3>=1.26.12 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
uvicorn>=0.20.0 ; python_version >= '3.7'
w3lib>=2.0.1 ; python_version >= '3.6'
wcwidth>=0.2.5
webencodings>=0.5.1
websocket-client>=1.4.2 ; python_version >= '3.7'
websockets>=10.4 ; python_version >= '3.7'
widgetsnbextension>=4.0.3 ; python_version >= '3.7'
yoyo-migrations>=8.1.0
zipp>=3.10.0 ; python_version >= '3.7'